Move eigenvalue computation to domainmatrix.py
